% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QHist.R
\name{q_hist}
\alias{q_hist}
\title{Creates a Histogram of a Continuous Variable}
\usage{
q_hist(x = rnorm(100), h_0 = 0)
}
\arguments{
\item{x}{a vector of numeric values (integer or real)}

\item{h_0}{a constant numeric for the expectation of the
null hypothesis. The default value is h_0 = 0.0.}
}
\value{
a plot of a histogram of the data, the mean and 95% confidence
interval and a one-sample t-test of the mean against the expectation of
the null hypothesis h_0. The default is h_0 = 0.0.
}
\description{
This function plots a histogram of a continuous variable,
along with its mean, 95% confidence interval, and a smoothed
density curve. A one-sample t-test compares the observed mean to
an expectation from the null hypothesis (default h_0 = 0.0)
}
\examples{
\dontrun{
q_hist(x=runif(1000),h_0 = 0.50)
}

}
